Add minimal Array proto dense array fast paths
Add minimal fast paths for Array.prototype operations like push() and pop(),
triggered when the 'this' binding is an Array which has its array part still
present.  This is the most common case and user code will expect for example
Array.prototype.push() to be relatively fast.

The main purpose of this commit is to figure out the necessary internal
helpers for fast paths.  More fast paths will be added separately.